--- seatunnel-core/seatunnel-flink-starter/src/main/java/org/apache/seatunnel/core/starter/SeatunnelFlink.java
@@ -1,36 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.seatunnel.core.starter;
-
-import org.apache.seatunnel.core.starter.args.FlinkCommandArgs;
-import org.apache.seatunnel.core.starter.command.Command;
-import org.apache.seatunnel.core.starter.command.FlinkCommandBuilder;
-import org.apache.seatunnel.core.starter.config.FlinkJobType;
-import org.apache.seatunnel.core.starter.exception.CommandException;
-import org.apache.seatunnel.core.starter.utils.CommandLineUtils;
-
-public class SeatunnelFlink {
-
-    public static void main(String[] args) throws CommandException {
-        FlinkCommandArgs flinkCommandArgs = CommandLineUtils.parseCommandArgs(args, FlinkJobType.JAR);
-        Command<FlinkCommandArgs> flinkCommand = new FlinkCommandBuilder()
-            .buildCommand(flinkCommandArgs);
-        Seatunnel.run(flinkCommand);
-    }
-
-}
